New Feature: Specific Item Focus
Based off alexanderabramov's suggestion, see https://github.com/Expugn/priconne-quest-helper/issues/12 and https://github.com/Expugn/priconne-quest-helper/pull/13

Clicking on an item in the items tab categories will "focus" it.
When focused, the item will be highlighted red in the items tab and only quests containing that item will appear.

Items can be unfocused by clicking on the focused item again or the focus can be switched if any other item is selected.

The item does not need to be "enabled/required" to be focused on.