Neneka and Equipment Cost Update
priconne has changed a lot of equipment recipes to make it easier for people to craft annoying stuff. Because of this change, it makes it awkward for people who don't play in the JP region since the requirements to craft stuff is different.

TW / KR PLAYERS: CHANGE `Equipment Data Type` to `Legacy Equipment Data (Before 2019.08.31) (TW / KR)`! USING THE CURRENT EQUIPMENT DATA IS INCORRECT FOR YOUR REGION!

- Added an option to switch to the Legacy Equipment Data in the quest helper/recipe data/statistics pages
- New Setting: Equipment Data Type
  - Can switch and save between current and legacy equipment data.
- New title background because of the new event.
- New text that needs to be translated
- Character data for Neneka
- Modified meta tags
- Fixed a bug where the add/sub/prioritize/deprioritize buttons would disable themselves when the language is changed; this should have not been the case.
- Removed usage of favicon.ico
- Clicking `Clear Item Tables` no longer wipes the starting tips.