Future-Proof Previous Hot-Fix and Other Changes
This is a continuation of the Hot-Fix pushed from the last commit.
The fix was hard-coded, so if a new character had the same issue the list would need to be updated manually.
This commit makes the hot-fix more future proof so that won't need to be done in the future.

- Adjusted text positioning for mobile
- Adjusted preset select list width for mobile
- Adjusted memory piece and memory piece drop-rate positioning for mobile
- Adjusted max device width for mobile before everything goes bad, aka 400px