Fix Issue
There was a problem where even if the inline inventory editor wasnt open the inventory button still remained disabled

Going with an alternate way of handling this by checking if there's any inline editor open if the inventory modal button is clicked. The cost of doing it this way however is that there's no visual indicator if it's available to be used or not.